Move method parent functions from `InheritedMethodWithNoCodeCheck` to `RoutineUtils`
